+++ src/hooks/chances/RandomMirror.cpp
@@ -52,7 +52,7 @@ class $modify(RandomMirrorPlayLayer, PlayLayer) {
 
     void nextFlipPortal(float) {
         log::trace("scheduling flip");
-        scheduleOnce(schedule_selector(RandomMirrorPlayLayer::flipPortal), rng::get(10.f, 1.f) * chanceToDelayPct(m_fields->chance));
+        scheduleOnce(schedule_selector(RandomMirrorPlayLayer::flipPortal), rng::get(10.f, 1.f) * rng::chanceToDelayPct(m_fields->chance));
     };
 
     void flipPortal(float) {